Tests for the Lion class are divided into two groups, parameterized and not. Checkin exception message is implemented via the assertThrows method.
