+++ src/main/java/com/example/Alex.java
@@ -3,7 +3,6 @@ package com.example;
 import java.util.List;
 
 public class Alex extends Lion {
-
     public Alex() throws Exception {
         super(new Feline(), "самец");
     }
